Split outputs and cleaned style
Changed the outputs to give one tidy csv of schools that matched and one tidy csv of schools that did not match. 

I also went through and tried to standardize my naming conventions somewhat. It isn't the cleanest, and I made overzealous use of underscores, but it is a little more logical now.